LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/geolocation-API/enabled-by-permission-policy-attribute-redirect-on-load.https.sub-expected.txt

CONSOLE MESSAGE: Blocked https://127.0.0.1:9443/resources/testdriver.js from asking for credentials because it is a cross-origin request.
CONSOLE MESSAGE: Permission policy 'Geolocation' check failed for document with origin 'https://127.0.0.1:9443'.


PASS Permissions-Policy allow="geolocation" allows same-origin redirection
FAIL Permissions-Policy allow="geolocation" allows cross-origin redirection assert_true: Geolocation API expected true got false
